Add extra credit progressive darkening effect
We want it to be completely black after 10 "iterations" so taking off
10% of the color each time wouldn't work (ex. r * 0.9, g * 0.9, b * 0.9)
since that just repeatedly take 10% off of the color so the rate at
which it gets to black decreases as it gets darker since the rgb is
closer to 0 and 10% of that small number is small. It doesn't actually
guarantee that we always get to black within 10 interactions.

What we actually need to do is subtract 10% off of white, the max
rgb value each time, this is taking off a 10% step each interaction
which means we will always get to black in 10 interactions
regardless of color (ex. r - 25.5 [25.5 is 10% of 255 white]...)